requirements.txt
torch
transformers
accelerate
bitsandbytes
sentencepiece
protobuf
huggingface_hub
llama-cpp-python
